fix clean_g2p method, move base_dataset and example_dataset to experiment directory
